got tests.py to work I also started on creating the form and routes for follow unfollow

+++ app/routes.py
@@ -8,7 +8,7 @@ from flask import render_template, flash, redirect, url_for, request
 from flask_login import login_user, logout_user, current_user, login_required
 import sqlalchemy as sa
 from app import app, db
-from app.forms import LoginForm, RegistrationForm, EditProfileForm
+from app.forms import LoginForm, RegistrationForm, EditProfileForm, EmptyForm
 from app.models import User
 
 @app.before_request
@@ -111,7 +111,8 @@ def edit_profile():
             current_user.image_file = picture_file
             db.session.commit()
             try:
-                os.remove(os.path.join(app.root_path, 'static/profile_pics', image_name))
+                if image_name != 'default.jpg':
+                    os.remove(os.path.join(app.root_path, 'static/profile_pics', image_name))
             except:
                 pass
             image_file = url_for('static', filename='profile_pics/' + current_user.image_file)
@@ -128,7 +129,48 @@ def edit_profile():
         form.username.data = current_user.username
         form.about_me.data = current_user.about_me
         return render_template('edit_profile.html', form=form, title='Edit Profile', image_file=image_file)
-     
+
+@app.route('/follow/<username>', methods=['POST'])
+@login_required
+def follow(username):
+    form = EmptyForm()
+    if form.validate_on_submit():
+        user = db.session.scalar(
+            sa.select(User).where(User.username == username))
+        if user is None:
+            flash(f'User {username} not found.')
+            return redirect(url_for('index'))
+        if user == current_user:
+            flash('You cannot follow yourself!')
+            return redirect(url_for('user', username=username))
+        current_user.follow(user)
+        db.session.commit()
+        flash(f'You are following {username}!')
+        return redirect(url_for('user', username=username))
+    else:
+        return redirect(url_for('index'))
+
+
+@app.route('/unfollow/<username>', methods=['POST'])
+@login_required
+def unfollow(username):
+    form = EmptyForm()
+    if form.validate_on_submit():
+        user = db.session.scalar(
+            sa.select(User).where(User.username == username))
+        if user is None:
+            flash(f'User {username} not found.')
+            return redirect(url_for('index'))
+        if user == current_user:
+            flash('You cannot unfollow yourself!')
+            return redirect(url_for('user', username=username))
+        current_user.unfollow(user)
+        db.session.commit()
+        flash(f'You are not following {username}.')
+        return redirect(url_for('user', username=username))
+    else:
+        return redirect(url_for('index'))
+
